Chore: refactor so contract doesn't depend on contract-tools
