+++ apps/mcp/tools/manifest/manifest-init.tool.ts
@@ -1,12 +1,14 @@
-import { z } from "zod";
-import { initManifest } from "@rejot-dev/contract-tools/manifest";
-import type { IRejotMcp, IFactory } from "@/rejot-mcp";
-import type { McpState } from "@/state/mcp-state";
 import { join } from "node:path";
-import { ensurePathRelative } from "@/util/fs.util";
+
 import { getLogger } from "@rejot-dev/contract/logger";
+import { initManifest } from "@rejot-dev/contract-tools/manifest";
 import { mergeAndUpdateManifest } from "@rejot-dev/contract-tools/manifest/manifest.fs";
-import type { IWorkspaceService } from "@rejot-dev/contract/workspace";
+import type { IWorkspaceService } from "@rejot-dev/contract-tools/manifest/manifest-workspace-resolver";
+import { z } from "zod";
+
+import type { IFactory, IRejotMcp } from "@/rejot-mcp";
+import type { McpState } from "@/state/mcp-state";
+import { ensurePathRelative } from "@/util/fs.util";
 
 const log = getLogger(import.meta.url);
